Adjust related quiz sections and footer tagline
